refactor(frontend): move SearchQuery to its own file
